third_party/blink/web_tests/platform/win11-arm64/external/wpt/uievents/interface/keyboard-accesskey-click-event-expected.txt

This is a testharness.js-based test.
[FAIL] Test that the Keyboard generated click does not fire pointer or mouse events for button
  assert_equals: Click's pointerId has default value expected 0 but got -1
[FAIL] Test that the Keyboard generated click does not fire pointer or mouse events for inputbutton
  assert_equals: Click's pointerId has default value expected 0 but got -1
[FAIL] Test that the Keyboard generated click does not fire pointer or mouse events for label
  assert_equals: Click's pointerId has default value expected 0 but got -1
Harness: the test ran to completion.
